+++ CHANGELOG.md
@@ -17,6 +17,7 @@ New features:
 - Display a link to the generated docs' `index.html` (#379)
 - Add `--open` flag to `spago docs` which opens generated docs in browser (#379)
 - Support building for alternate backends (#355). E.g: Use `backend = "psgo"` entry in `spago.dhall` to compile with `psgo`
+- Add `--no-comments` flag to `spago init` which strips comments from the generated `spago.dhall` and `packages.dhall` configs (#417)
 
 Bugfixes:
 - Warn (but don't error) when trying to watch missing directories (#406)